Replace Deck implementation
HTMLElements that were pre-written are replaced by dynamically generated
HTMLElements based on javascript objects. Each Card object contains a
private field with the HTML element for that Card. The HTML element will
only need to be created the first time the Card is used and will be
reused any subsequent times it is played.

By dynamically creating the Cards, the number of decks in play can be
more easily set by simply changing the value of const kNumDecks.

Dealt cards are placed in a separated discard array so
that the entire deck will be used before being reshuffled back into
play.

Added text for Surrender instead of showing Lose text.