V4.0 Add Splitting
Splitting now creates another playable hand for the player. When
splitting, the player will play each hand in sequence and then will be
prompted before showing the final results for each hand.

Replace constants with immutable objects.

Deck implementation - Old implementation had an array of 52 elements
with a counter based on the number of decks. Now, each card is its own
element so with multiple decks, there are multiple array entries of the
same card.